requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==27.0.0
flake8==7.1.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
kombu==5.4.2
mccabe==0.7.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pyflakes==3.2.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
